Adding a server that handles multiple clients connected to it. This will also echo messages to all clients that are connected if entered on the server console and start with the characters ::
So the following line will send a message to all connected clients.
::Hi clients!

Telnet clients will get this instantly. The python clients need to be modified to listen for input on a thread so it isn't blocked by keyboard input. This won't really matter on the real client as there won't be keyboard entry on those clients.

Code initially came from the following locations:
    http://ilab.cs.byu.edu/python/threadingmodule.html
    http://ilab.cs.byu.edu/python/code/echoclient-select.py